Removed the frequent itemset package with its classes.
I am not an expert in frequent item set mining and I would need a lot
more time to become one. I want to concentrate on clustering algorithms
for now and therefore, the frequent item set mining part is excluded
from the package.

Maybe the package reappears some time in the future, but first I need
nore knowledge in the field and I need a useful visualization which is
not available at the moment.