CLOSED - task 8: Javadoc comments for all classes in the etc package http://github.com/Roland-Winkler/EDMOAL/issues/issue/8
I removed almost all functions from the MyMath class as they did not
work well with the general applicability of the remaining project. Most
of the functions were distance functions anyway which are available in
DAEuclideanMetric.
The other functions were too specialised to be in the MyMath class. 

As a consequence, the double array related sequence classes had to be
updated as well because they used the now removed distance functions. I
also corrected theses errors.