CLOSED - task 24: Make copy constructors of clustering algorithms use the abstract method to be able to initialize a new clustering algorithm with a different one. http://github.com/Roland-Winkler/EDMOAL/issues/issue/24
also some of the commentary is already in the hard c-means clustering
algorithm class. But thats relatively small amount and removing the
inconsistency was more important than providing the comments. Especially
because the comments for the now changed functions are different to the
ones in the old version. win/win :)